[No-ticket] Split into 2 files
